feat: move config into package
